format-patch: allow a non-integral version numbers
Usually we can only use `format-patch -v<n>` to generate integral
version numbers patches, but sometimes a same fixup should be
labeled as a non-integral versions like `v1.1`, so teach `format-patch`
to allow a non-integral versions may be helpful to send those patches.

Since the original `format-patch` logic, if we specify a version `-v<n>`
and commbine with `--interdiff` or `--rangediff`, the patch will output
"Interdiff again v<n-1>:" or "Range-diff again v<n-1>:`, but this does
not meet the requirements of our fractional version numbers, so if the
user use a integral version number `-v<n>`, ensure that the output in
the patch is still `v<n-1>`; otherwise, only output "Interdiff" or
"Range-diff".

Signed-off-by: ZheNing Hu <adlternative@gmail.com>