Documentation: amended usages of various (sub)commands
The Git suite option parsing API's Technical Documentation suggests this
about the option descriptions of a (sub)command:

`description` is a short string to describe the effect of the option.
It shall begin with a lower-case letter and a full stop (.) shall be
omitted at the end.

Various (sub)commands' option arrays don't follow the guideline provided
by the parse_options Documentation regarding the descriptions.

Signed-off-by: Chinmoy Chakraborty <chinmoy12c@gmail.com>